+++ src/java/org/malbino/orion/controllers/AbstractController.java
@@ -22,6 +22,7 @@ import org.malbino.orion.entities.Estudiante;
 import org.malbino.orion.entities.GestionAcademica;
 import org.malbino.orion.entities.Nota;
 import org.malbino.orion.entities.Postulante;
+import org.malbino.orion.entities.Usuario;
 import org.malbino.orion.enums.Caracter;
 import org.malbino.orion.enums.Condicion;
 import org.malbino.orion.enums.Funcionalidad;
@@ -38,6 +39,7 @@ import org.malbino.orion.facades.EmpleadoFacade;
 import org.malbino.orion.facades.EstudianteFacade;
 import org.malbino.orion.facades.GestionAcademicaFacade;
 import org.malbino.orion.facades.PostulanteFacade;
+import org.malbino.orion.facades.UsuarioFacade;
 import org.malbino.orion.util.Redondeo;
 import org.primefaces.PrimeFaces;
 
@@ -59,6 +61,8 @@ public abstract class AbstractController implements Serializable {
     GestionAcademicaFacade gestionAcademicaFacade;
     @EJB
     PostulanteFacade postulanteFacade;
+    @EJB
+    UsuarioFacade usuarioFacade;
 
     protected void mensajeDeError(String mensaje) {
         FacesContext context = FacesContext.getCurrentInstance();
@@ -162,6 +166,19 @@ public abstract class AbstractController implements Serializable {
         }
 
         return postulantesFiltrados;
+    }
+    
+    public List<Usuario> completarUsuario(String consulta) {
+        List<Usuario> usuarios = usuarioFacade.findAll();
+        List<Usuario> usuariosFiltrados = new ArrayList();
+
+        for (Usuario u : usuarios) {
+            if (u.toString().toLowerCase().contains(consulta.toLowerCase())) {
+                usuariosFiltrados.add(u);
+            }
+        }
+
+        return usuariosFiltrados;
     }
 
     public List<Campus> listaCampus() {