+++ src/java/org/malbino/orion/controllers/AdmisionesController.java
@@ -5,28 +5,17 @@
 package org.malbino.orion.controllers;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.Serializable;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.util.Date;
+import java.util.ArrayList;
+import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.ejb.EJB;
 import javax.enterprise.context.SessionScoped;
 import javax.inject.Named;
-import org.apache.commons.io.FilenameUtils;
 import org.malbino.orion.entities.Carrera;
 import org.malbino.orion.entities.GestionAcademica;
 import org.malbino.orion.entities.Postulante;
-import org.malbino.orion.enums.Funcionalidad;
-import org.malbino.orion.facades.ActividadFacade;
-import org.malbino.orion.facades.PagoFacade;
 import org.malbino.orion.facades.PostulanteFacade;
-import org.malbino.orion.facades.negocio.AdmisionesFacade;
-import org.malbino.orion.util.Fecha;
-import org.primefaces.event.FileUploadEvent;
 
 /**
  *
@@ -37,179 +26,130 @@ import org.primefaces.event.FileUploadEvent;
 public class AdmisionesController extends AbstractController implements Serializable {
 
     @EJB
-    ActividadFacade actividadFacade;
-    @EJB
     PostulanteFacade postulanteFacade;
-    @EJB
-    AdmisionesFacade admisionesFacade;
-    @EJB
-    PagoFacade pagoFacade;
 
-    private Postulante nuevoPostulante;
+    private GestionAcademica seleccionGestionAcademica;
+    private Carrera seleccionCarrera;
+
+    private List<Postulante> postulantes;
     private Postulante seleccionPostulante;
 
-    private String ci;
-    private Date fechaNacimiento;
-    private GestionAcademica gestionAcademica;
-    private Carrera carrera;
+    private Boolean filter;
+    private String keyword;
 
     @PostConstruct
     public void init() {
-        nuevoPostulante = new Postulante();
+        seleccionGestionAcademica = null;
+        seleccionCarrera = null;
+
+        postulantes = new ArrayList<>();
         seleccionPostulante = null;
 
-        ci = null;
-        fechaNacimiento = null;
-        gestionAcademica = null;
-        carrera = null;
+        filter = false;
+        keyword = null;
     }
 
     public void reinit() {
-        nuevoPostulante = new Postulante();
+        if (seleccionGestionAcademica != null && seleccionCarrera == null) {
+            postulantes = postulanteFacade.listaPostulantes(seleccionGestionAcademica.getId_gestionacademica());
+        } else if (seleccionGestionAcademica != null && seleccionCarrera != null) {
+            postulantes = postulanteFacade.listaPostulantes(seleccionGestionAcademica.getId_gestionacademica(), seleccionCarrera.getId_carrera());
+        }
         seleccionPostulante = null;
 
-        ci = null;
-        fechaNacimiento = null;
-        gestionAcademica = null;
-        carrera = null;
+        filter = false;
+        keyword = null;
     }
 
-    public void subirArchivoRegistrarPostulante(FileUploadEvent event) {
-        Path folder = Paths.get(realPath() + "/resources/uploads/photos");
-        String extension = FilenameUtils.getExtension(event.getFile().getFileName());
-        Path file = null;
-        try (InputStream input = event.getFile().getInputStream()) {
-            file = Files.createTempFile(folder, null, "." + extension);
-            Files.copy(input, file, StandardCopyOption.REPLACE_EXISTING);
-
-            nuevoPostulante.setFoto(file.getFileName().toString());
-        } catch (IOException ex) {
-
+    @Override
+    public List<Carrera> listaCarreras() {
+        List<Carrera> l = new ArrayList();
+        if (seleccionGestionAcademica != null) {
+            l = carreraFacade.listaCarreras(seleccionGestionAcademica.getRegimen());
         }
+        return l;
     }
 
-    public void subirArchivoModificarPostulante(FileUploadEvent event) {
-        Path folder = Paths.get(realPath() + "/resources/uploads/photos");
-        String extension = FilenameUtils.getExtension(event.getFile().getFileName());
-        Path file = null;
-        try (InputStream input = event.getFile().getInputStream()) {
-            file = Files.createTempFile(folder, null, "." + extension);
-            Files.copy(input, file, StandardCopyOption.REPLACE_EXISTING);
+    public void filtro() {
+        if (filter) {
+            filter = false;
+            keyword = null;
 
-            seleccionPostulante.setFoto(file.getFileName().toString());
-        } catch (IOException ex) {
-
-        }
-    }
-
-    public void registrarPostulante() throws IOException {
-        if (!actividadFacade.listaActividades(Fecha.getDate(), Funcionalidad.ADMISION, nuevoPostulante.getGestionAcademica().getId_gestionacademica()).isEmpty()) {
-            if (postulanteFacade.buscarPostulante(nuevoPostulante.getCi(), nuevoPostulante.getGestionAcademica().getId_gestionacademica(), nuevoPostulante.getCarrera().getId_carrera()) == null) {
-                if (nuevoPostulante.getFoto() != null && !nuevoPostulante.getFoto().isEmpty()) {
-                    if (admisionesFacade.registrarPostulante(nuevoPostulante)) {
-                        reinit();
-
-                        this.mensajeDeInformacion("Postulante registrado exitosamente.");
-                    } else {
-                        this.mensajeDeError("No se pudo registrar al postulante.");
-                    }
-                } else {
-                    this.mensajeDeError("* Fotografía requerida.");
-                }
-            } else {
-                this.mensajeDeError("Postulante repetido.");
+            if (seleccionGestionAcademica != null && seleccionCarrera == null) {
+                postulantes = postulanteFacade.listaPostulantes(seleccionGestionAcademica.getId_gestionacademica());
+            } else if (seleccionGestionAcademica != null && seleccionCarrera != null) {
+                postulantes = postulanteFacade.listaPostulantes(seleccionGestionAcademica.getId_gestionacademica(), seleccionCarrera.getId_carrera());
             }
         } else {
-            this.mensajeDeError("Fuera de fecha.");
+            filter = true;
+            keyword = null;
         }
     }
 
-    public void buscarPostulanteModificarPostulante() throws IOException {
-        if (!actividadFacade.listaActividades(Fecha.getDate(), Funcionalidad.ADMISION, gestionAcademica.getId_gestionacademica()).isEmpty()) {
-            seleccionPostulante = postulanteFacade.buscarPostulante(ci, fechaNacimiento, gestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
-            if (seleccionPostulante != null) {
-                if (!pagoFacade.listaPagosAdeudadosPostulante(seleccionPostulante.getId_postulante()).isEmpty()) {
-                    toModificarPostulante();
-                } else {
-                    this.mensajeDeError("Postulante inscrito.");
-                }
-            } else {
-                this.mensajeDeError("Postulante no encontrado.");
-            }
-        } else {
-            this.mensajeDeError("Fuera de fecha.");
+    public void buscar() {
+        if (seleccionGestionAcademica != null && seleccionCarrera == null) {
+            postulantes = postulanteFacade.buscar(seleccionGestionAcademica.getId_gestionacademica(), keyword);
+        } else if (seleccionGestionAcademica != null && seleccionCarrera != null) {
+            postulantes = postulanteFacade.buscar(seleccionGestionAcademica.getId_gestionacademica(), seleccionCarrera.getId_carrera(), keyword);
         }
     }
 
-    public void buscarPostulanteImprimir() throws IOException {
-        if (!actividadFacade.listaActividades(Fecha.getDate(), Funcionalidad.ADMISION, gestionAcademica.getId_gestionacademica()).isEmpty()) {
-            seleccionPostulante = postulanteFacade.buscarPostulante(ci, fechaNacimiento, gestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
-            if (seleccionPostulante != null) {
-                if (!pagoFacade.listaPagosAdeudadosPostulante(seleccionPostulante.getId_postulante()).isEmpty()) {
-                    this.insertarParametro("id_postulante", seleccionPostulante.getId_postulante());
+    public void imprimirFormularioPostulante() throws IOException {
+        if (seleccionPostulante != null) {
+            this.insertarParametro("id_postulante", seleccionPostulante.getId_postulante());
 
-                    toFormularioPostulante();
-                } else {
-                    this.mensajeDeError("Postulante inscrito.");
-                }
-            } else {
-                this.mensajeDeError("Postulante no encontrado.");
-            }
-        } else {
-            this.mensajeDeError("Fuera de fecha.");
+            toFormularioPostulante();
         }
-    }
-
-    public void modificarPostulante() throws IOException {
-        if (!actividadFacade.listaActividades(Fecha.getDate(), Funcionalidad.ADMISION, seleccionPostulante.getGestionAcademica().getId_gestionacademica()).isEmpty()) {
-            if (postulanteFacade.buscarPostulante(seleccionPostulante.getCi(), seleccionPostulante.getGestionAcademica().getId_gestionacademica(), seleccionPostulante.getCarrera().getId_carrera(), seleccionPostulante.getId_postulante()) == null) {
-                if (seleccionPostulante.getFoto() != null && !seleccionPostulante.getFoto().isEmpty()) {
-                    if (postulanteFacade.edit(seleccionPostulante)) {
-                        this.mensajeDeInformacion("Postulante modificado exitosamente.");
-                    } else {
-                        this.mensajeDeError("No se pudo modificar al postulante.");
-                    }
-                } else {
-                    this.mensajeDeError("* Fotografía requerida.");
-                }
-            } else {
-                this.mensajeDeError("Postulante repetido.");
-            }
-        } else {
-            this.mensajeDeError("Fuera de fecha.");
-        }
-    }
-
-    public void toRegistrarPostulante() throws IOException {
-        this.redireccionarViewId("/admisiones/registrarPostulante");
-    }
-
-    public void toModificarPostulante() throws IOException {
-        this.redireccionarViewId("/admisiones/modificar/modificarPostulante");
     }
 
     public void toFormularioPostulante() throws IOException {
-        this.redireccionarViewId("/admisiones/imprimir/formularioPostulante");
+        this.redireccionarViewId("/admisiones/formularioPostulante");
     }
 
-    public void toHome() throws IOException {
-        this.reinit();
-
-        this.redireccionarViewId("/admisiones/home");
-    }
-
-    /**
-     * @return the nuevoPostulante
-     */
-    public Postulante getNuevoPostulante() {
-        return nuevoPostulante;
+    public void toPostulantes() throws IOException {
+        this.redireccionarViewId("/admisiones/postulantes");
     }
 
     /**
-     * @param nuevoPostulante the nuevoPostulante to set
+     * @return the seleccionGestionAcademica
      */
-    public void setNuevoPostulante(Postulante nuevoPostulante) {
-        this.nuevoPostulante = nuevoPostulante;
+    public GestionAcademica getSeleccionGestionAcademica() {
+        return seleccionGestionAcademica;
+    }
+
+    /**
+     * @param seleccionGestionAcademica the seleccionGestionAcademica to set
+     */
+    public void setSeleccionGestionAcademica(GestionAcademica seleccionGestionAcademica) {
+        this.seleccionGestionAcademica = seleccionGestionAcademica;
+    }
+
+    /**
+     * @return the seleccionCarrera
+     */
+    public Carrera getSeleccionCarrera() {
+        return seleccionCarrera;
+    }
+
+    /**
+     * @param seleccionCarrera the seleccionCarrera to set
+     */
+    public void setSeleccionCarrera(Carrera seleccionCarrera) {
+        this.seleccionCarrera = seleccionCarrera;
+    }
+
+    /**
+     * @return the postulantes
+     */
+    public List<Postulante> getPostulantes() {
+        return postulantes;
+    }
+
+    /**
+     * @param postulantes the postulantes to set
+     */
+    public void setPostulantes(List<Postulante> postulantes) {
+        this.postulantes = postulantes;
     }
 
     /**
@@ -227,58 +167,31 @@ public class AdmisionesController extends AbstractController implements Serializ
     }
 
     /**
-     * @return the ci
+     * @return the filter
      */
-    public String getCi() {
-        return ci;
+    public Boolean getFilter() {
+        return filter;
     }
 
     /**
-     * @param ci the ci to set
+     * @param filter the filter to set
      */
-    public void setCi(String ci) {
-        this.ci = ci;
+    public void setFilter(Boolean filter) {
+        this.filter = filter;
     }
 
     /**
-     * @return the fechaNacimiento
+     * @return the keyword
      */
-    public Date getFechaNacimiento() {
-        return fechaNacimiento;
+    public String getKeyword() {
+        return keyword;
     }
 
     /**
-     * @param fechaNacimiento the fechaNacimiento to set
+     * @param keyword the keyword to set
      */
-    public void setFechaNacimiento(Date fechaNacimiento) {
-        this.fechaNacimiento = fechaNacimiento;
+    public void setKeyword(String keyword) {
+        this.keyword = keyword;
     }
 
-    /**
-     * @return the gestionAcademica
-     */
-    public GestionAcademica getGestionAcademica() {
-        return gestionAcademica;
-    }
-
-    /**
-     * @param gestionAcademica the gestionAcademica to set
-     */
-    public void setGestionAcademica(GestionAcademica gestionAcademica) {
-        this.gestionAcademica = gestionAcademica;
-    }
-
-    /**
-     * @return the carrera
-     */
-    public Carrera getCarrera() {
-        return carrera;
-    }
-
-    /**
-     * @param carrera the carrera to set
-     */
-    public void setCarrera(Carrera carrera) {
-        this.carrera = carrera;
-    }
 }